+++ groucho-lsh-rest/src/main/java/it/cnr/iasi/saks/groucho/lsh/rest/api/impl/MarkApiController.java
@@ -1,7 +1,8 @@
-package it.cnr.iasi.saks.groucho.lsh.rest.api;
+package it.cnr.iasi.saks.groucho.lsh.rest.api.impl;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import io.swagger.annotations.ApiParam;
+import it.cnr.iasi.saks.groucho.lsh.rest.api.MarkApi;
 import it.cnr.iasi.saks.groucho.lsh.service.MarkApiService;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.*;
